requirements.txt
google-cloud-bigquery==3.10.0
Faker==18.6.2
pandas==2.0.1
db-dtypes == 1.1.1